fix double types

+++ src/main/java/es/uma/informatica/sii/entities/Expediente.java
@@ -17,104 +17,117 @@ import javax.persistence.OneToMany;
 @Entity
 
 public class Expediente implements Serializable {
- 
+
 	@Id
-	@Column(length=9)
+	@Column(length = 9)
 	private Integer numero;
 	private Boolean activo;
 	private Double nota_media;
-	private Integer creditos_fb;
-	private Integer creditos_ob;
-	private Integer creditos_pe;
-	private Integer creditos_tf;
-	
+	private Double creditos_fb;
+	private Double creditos_ob;
+	private Double creditos_pe;
+	private Double creditos_tf;
+
 	@ManyToOne
-	@JoinColumn(nullable=false)
+	@JoinColumn(nullable = false)
 	private Titulacion titulacion;
-	
+
 	@ManyToOne
-	@JoinColumn(nullable=false)
+	@JoinColumn(nullable = false)
 	private Alumno alumno;
-	
-	@OneToMany(mappedBy="expediente")
+
+	@OneToMany(mappedBy = "expediente")
 	private List<Encuesta> encuestas;
-	
-	@OneToMany(mappedBy="expediente")
+
+	@OneToMany(mappedBy = "expediente")
 	private List<Matricula> matriculas;
-	
+
 	private static final long serialVersionUID = 1L;
 
 	public Expediente() {
 		super();
-	}   
+	}
+
 	public Integer getNumero() {
 		return this.numero;
 	}
 
 	public void setNumero(Integer numero) {
 		this.numero = numero;
-	}   
+	}
+
 	public Boolean getActivo() {
 		return this.activo;
 	}
 
 	public void setActivo(Boolean activo) {
 		this.activo = activo;
-	}   
+	}
+
 	public Double getNota_media() {
 		return this.nota_media;
 	}
 
 	public void setNota_media(Double nota_media) {
 		this.nota_media = nota_media;
-	}   
-	public Integer getCreditos_fb() {
+	}
+
+	public Double getCreditos_fb() {
 		return this.creditos_fb;
 	}
 
-	public void setCreditos_fb(Integer creditos_fb) {
+	public void setCreditos_fb(Double creditos_fb) {
 		this.creditos_fb = creditos_fb;
-	}   
+	}
+
 	public Titulacion getTitulacion() {
 		return titulacion;
 	}
+
 	public void setTitulacion(Titulacion titulacion) {
 		this.titulacion = titulacion;
 	}
+
 	public Alumno getAlumno() {
 		return alumno;
 	}
+
 	public void setAlumno(Alumno alumno) {
 		this.alumno = alumno;
 	}
-	public Integer getCreditos_ob() {
+
+	public Double getCreditos_ob() {
 		return this.creditos_ob;
 	}
 
-	public void setCreditos_ob(Integer creditos_ob) {
+	public void setCreditos_ob(Double creditos_ob) {
 		this.creditos_ob = creditos_ob;
-	}   
-	public Integer getCreditos_pe() {
+	}
+
+	public Double getCreditos_pe() {
 		return this.creditos_pe;
 	}
 
-	public void setCreditos_pe(Integer creditos_pe) {
+	public void setCreditos_pe(Double creditos_pe) {
 		this.creditos_pe = creditos_pe;
-	}   
-	public Integer getCreditos_tf() {
+	}
+
+	public Double getCreditos_tf() {
 		return this.creditos_tf;
 	}
 
-	public void setCreditos_tf(Integer creditos_tf) {
+	public void setCreditos_tf(Double creditos_tf) {
 		this.creditos_tf = creditos_tf;
-	}   
+	}
+
 	public Titulacion getTitulacion_codigo() {
 		return this.titulacion;
 	}
 
 	public void setTitulacion_codigo(Titulacion titulacion_codigo) {
 		this.titulacion = titulacion_codigo;
-	}   
+	}
+
 	public Alumno getAlumno_dni() {
 		return this.alumno;
 	}
@@ -122,20 +135,23 @@ public class Expediente implements Serializable {
 	public void setAlumno_dni(Alumno alumno_dni) {
 		this.alumno = alumno_dni;
 	}
-	
+
 	public List<Encuesta> getEncuestas() {
 		return encuestas;
 	}
+
 	public void setEncuestas(List<Encuesta> encuestas) {
 		this.encuestas = encuestas;
 	}
+
 	public List<Matricula> getMatriculas() {
 		return matriculas;
 	}
+
 	public void setMatriculas(List<Matricula> matriculas) {
 		this.matriculas = matriculas;
 	}
-	
+
 	@Override
 	public int hashCode() {
 		final int prime = 31;
@@ -143,6 +159,7 @@ public class Expediente implements Serializable {
 		result = prime * result + ((numero == null) ? 0 : numero.hashCode());
 		return result;
 	}
+
 	@Override
 	public boolean equals(Object obj) {
 		if (this == obj)
@@ -157,9 +174,10 @@ public class Expediente implements Serializable {
 			return false;
 		return true;
 	}
+
 	@Override
 	public String toString() {
 		return "Expediente [numero=" + numero + ", alumno=" + alumno + "]";
 	}
-   
+
 }
